+++ nsgt/fscale.py
@@ -14,7 +14,8 @@ AudioMiner project, supported by Vienna Science and Technology Fund (WWTF)
 import numpy as np
 from nsgt.utilities.compat import xrange
 
-class Scale:
+
+class Scale(object):
     dbnd = 1.e-8
 
     def __init__(self, bnds):
