+++ nsgt/slicing.py
@@ -21,7 +21,8 @@ from nsgt.utilities.compat import izip, xrange
 def makewnd(sl_len, tr_area):
     hhop = sl_len // 4
     htr = tr_area // 2
-    # build window function within one slice (centered with transition areas around sl_len/4 and 3*sl_len/4    
+    # build window function within one slice (centered with
+    # transition areas around sl_len/4 and 3*sl_len/4
     w = hannwin(2 * tr_area)  # window is shifted
     tw = np.empty(sl_len, dtype=float)
     tw[:hhop - htr] = 0
@@ -45,20 +46,22 @@ def slicing(f, sl_len, tr_area):
     tw = [tw[o:o + hhop] for o in xrange(0, sl_len, hhop)]
 
     # stream of hopsize/2 blocks with leading and trailing zero blocks
-    fseq = reblock(f, hhop, dtype=float, fulllast=True, padding=0., multichannel=True)
+    fseq = reblock(f, hhop, dtype=float, fulllast=True,
+                   padding=0., multichannel=True)
 
     # get first block to deduce number of channels
     fseq0 = fseq.next()
     chns = len(fseq0)
     pad = np.zeros((chns, hhop), dtype=fseq0.dtype)
-    # assemble a stream of front padding, already retrieved first block, the block stream and some tail padding
+    # assemble a stream of front padding, already retrieved first block,
+    # the block stream and some tail padding
     fseq = chain((pad, pad, fseq0), fseq, (pad, pad, pad))
 
-    slices = [[slice(hhop * ((i + 3 - k * 2) % 4), hhop * ((i + 3 - k * 2) % 4 + 1)) for i in range(4)] for k in
-              range(2)]
+    slices = [[slice(hhop * ((i + 3 - k * 2) % 4), hhop * ((i + 3 - k * 2) % 4 + 1))
+               for i in range(4)] for k in range(2)]
     slices = cycle(slices)
 
-    past = []
+    past = list()
     for fi in fseq:
         past.append(fi)
         if len(past) == 4:
